WIP: split pde definitions from case definitions
