+++ timestep/include/ModelEvaluatorTPETRA_def.hpp
@@ -3013,7 +3013,7 @@ void ModelEvaluatorTPETRA<scalar_type>::set_test_case()
 
     paramfunc_.resize(2);
     paramfunc_[0] = &tools::solvers::param;
-    paramfunc_[1] = &cases::tonks1::param;
+    paramfunc_[1] = &pdes::kks::param;
 
     //post_proc.push_back(new post_process(mesh_,(int)0));
     //post_proc[0].postprocfunc_ = &tpetra::tonks::postproc_mu_a_;
